[Silver IV] Title: solved.ac, Time: 308 ms, Memory: 38052 KB -BaekjoonHub

+++ 백준/Silver/18110. solved．ac/solved．ac.java
@@ -8,7 +8,7 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
         int N = Integer.parseInt(br.readLine());
-        int cut = (int) Math.round((float) N * 0.15);
+        int cut = (int) Math.round((double) N * 0.15);
 
         int[] score = new int[N];
         for (int i = 0; i < N; i++) {
